1.6.0 release branch. Includes fetch refactor to support extensible transports and other enhancements, bugfixes.

+++ bdbag/bdbag_api.py
@@ -59,10 +59,7 @@ def read_metadata(metadata_file):
 def cleanup_bag(bag_path, save=False):
     logger.info("Cleaning up bag dir: %s" % bag_path)
     if save:
-        saved_bag_path = ''.join([bag_path, '_', datetime.strftime(datetime.now(), "%Y-%m-%d_%H.%M.%S")])
-        logger.info("Moving bag %s to %s" % (bag_path, saved_bag_path))
-        shutil.move(bag_path, saved_bag_path)
-        return saved_bag_path
+        return safe_move(bag_path)
     else:
         shutil.rmtree(bag_path)
         return None
@@ -81,7 +78,7 @@ def ensure_bag_path_exists(bag_path, save=True):
 
 def revert_bag(bag_path):
     if not is_bag(bag_path):
-        logger.warning("Cannot revert the bag %s because it is not a bag directory!")
+        logger.warning("Cannot revert the bag %s because it is not a bag directory!" % bag_path)
         return
 
     for path in os.listdir(bag_path):
@@ -90,12 +87,15 @@ def revert_bag(bag_path):
                 os.remove(os.path.join(bag_path, path))
 
     data_path = os.path.join(bag_path, 'data')
-    for path in os.listdir(data_path):
-        old_path = os.path.join(data_path, path)
-        new_path = os.path.join(bag_path, path)
-        logger.debug("Bag revert: moving payload file %s to %s", old_path, new_path)
-        os.rename(old_path, new_path)
-    os.rmdir(data_path)
+    if os.path.isdir(data_path):
+        for path in os.listdir(data_path):
+            old_path = os.path.join(data_path, path)
+            new_path = os.path.join(bag_path, path)
+            logger.debug("Bag revert: moving payload file %s to %s", old_path, new_path)
+            os.rename(old_path, new_path)
+        os.rmdir(data_path)
+    else:
+        logger.warning("Bag directory %s does not contain a \"data\" directory to revert." % bag_path)
     logging.info("Bag directory %s has been reverted back to a normal directory." % bag_path)
 
 
@@ -227,7 +227,7 @@ def make_bag(bag_path,
              metadata=None,
              metadata_file=None,
              remote_file_manifest=None,
-             config_file=DEFAULT_CONFIG_FILE,
+             config_file=None,
              ro_metadata=None,
              ro_metadata_file=None):
     bag = None
@@ -358,40 +358,50 @@ def extract_bag(bag_path, output_path=None, temp=False):
     if not os.path.exists(bag_path):
         raise RuntimeError("Specified bag path not found: %s" % bag_path)
 
+    # determine output path for extraction
+    extracted_path = None
     bag_dir = os.path.splitext(os.path.basename(bag_path))[0]
     if os.path.isfile(bag_path):
         if temp:
             output_path = tempfile.mkdtemp(prefix='bag_')
+        elif output_path:
+            safe_move(output_path, output_path)
         elif not output_path:
             output_path = os.path.splitext(bag_path)[0]
-            if os.path.exists(output_path):
-                newpath = ''.join([output_path, '-', datetime.strftime(datetime.now(), "%Y-%m-%d_%H.%M.%S")])
-                logger.info("Specified output path %s already exists, moving existing directory to %s" %
-                            (output_path, newpath))
-                shutil.move(output_path, newpath)
+            safe_move(output_path)
             output_path = os.path.dirname(bag_path)
+
+        # perform the extraction
         if zipfile.is_zipfile(bag_path):
             logger.info("Extracting ZIP archived file: %s" % bag_path)
             with open(bag_path, 'rb') as bag_file:
                 zipped = zipfile.ZipFile(bag_file)
+                files = zipped.namelist()
+                extracted_path = bag_parent_dir_from_archive(files)
                 zipped.extractall(output_path)
                 zipped.close()
         elif tarfile.is_tarfile(bag_path):
             logger.info("Extracting TAR/GZ/BZ2 archived file: %s" % bag_path)
             tarred = tarfile.open(bag_path)
+            files = tarred.getnames()
+            extracted_path = bag_parent_dir_from_archive(files)
             tarred.extractall(output_path)
             tarred.close()
         else:
             raise RuntimeError("Archive format not supported for file: %s"
                                "\nSupported archive formats are ZIP or TAR/GZ/BZ2" % bag_path)
 
-    extracted_path = os.path.join(output_path, bag_dir)
+    if not extracted_path:
+        extracted_path = os.path.join(output_path, bag_dir)
+    else:
+        extracted_path = os.path.join(output_path, extracted_path)
+
     logger.info("File %s was successfully extracted to directory %s" % (bag_path, extracted_path))
 
     return extracted_path
 
 
-def validate_bag(bag_path, fast=False, callback=None, config_file=DEFAULT_CONFIG_FILE):
+def validate_bag(bag_path, fast=False, callback=None, config_file=None):
     config = read_config(config_file)
     bag_config = config['bag_config']
     bag_processes = bag_config.get('bag_processes', 1)
@@ -409,7 +419,7 @@ def validate_bag(bag_path, fast=False, callback=None, config_file=DEFAULT_CONFIG
     except bdbagit.BaggingInterruptedError as e:
         logger.warning(get_typed_exception(e))
         raise e
-    except Exception as e:
+    except Exception as e:  # pragma: no cover
         raise RuntimeError("Unhandled exception while validating bag: %s" % e)
 
 
@@ -504,7 +514,7 @@ def generate_remote_files_from_manifest(remote_file_manifest, algs, strict=False
     return remote_files
 
 
-def generate_ro_manifest(bag_path, overwrite=False, config_file=DEFAULT_CONFIG_FILE):
+def generate_ro_manifest(bag_path, overwrite=False, config_file=None):
     bag = bdbagit.BDBag(bag_path)
     bag_ro_metadata_path = os.path.abspath(os.path.join(bag_path, "metadata", "manifest.json"))
     exists = os.path.isfile(bag_ro_metadata_path)
@@ -551,7 +561,7 @@ def resolve_fetch(bag_path,
                   force=False,
                   callback=None,
                   keychain_file=DEFAULT_KEYCHAIN_FILE,
-                  config_file=DEFAULT_CONFIG_FILE,
+                  config_file=None,
                   filter_expr=None,
                   **kwargs):
     bag = bdbagit.BDBag(bag_path)
@@ -576,7 +586,7 @@ def materialize(input_path,
                 fetch_callback=None,
                 validation_callback=None,
                 keychain_file=DEFAULT_KEYCHAIN_FILE,
-                config_file=DEFAULT_CONFIG_FILE,
+                config_file=None,
                 filter_expr=None,
                 force=False,
                 **kwargs):
@@ -598,7 +608,7 @@ def materialize(input_path,
             raise RuntimeError("Unable to retrieve bag from: %s" % input_path)
 
     if bag_file:
-        bag_path = extract_bag(bag_file)
+        bag_path = extract_bag(bag_file, output_path)
 
     if bag_path:
         if not is_bag(bag_path):
